1090_와 나... 범위 오류 나오네요... long int로 고칩니다.

+++ main.cpp
@@ -1,13 +1,13 @@
 #include <iostream>
 
 int main(){
-    int a,b,s,i; 
+    long int a,b,s,i; 
     
-    scanf("%d%d%d",&a,&b,&s);
+    scanf("%ld%ld%ld",&a,&b,&s);
     for(i=1;i<s;i++){////for 반목문에서 가운데의 조건이 빠진 경우 무한 반복된다.
       a*=b;
       
       //참이면 건너 띄고 반복함
       
-    }printf("%d ",a);
+    }printf("%ld ",a);
 }
